+++ src/HotChocolate/Core/src/Abstractions/Execution/Tasks/ExecutionTaskExtensions.cs
@@ -1,0 +1,14 @@
+namespace HotChocolate.Execution
+{
+    /// <summary>
+    /// Extensions for <see cref="IExecutionTask"/>.
+    /// </summary>
+    public static class ExecutionTaskExtensions
+    {
+        /// <summary>
+        /// Defines if this task is completed.
+        /// </summary>
+        public static bool IsCompleted(this IExecutionTask task)
+            => task.Status is ExecutionTaskStatus.Completed or ExecutionTaskStatus.Faulted;
+    }
+}
